fix(ci): address integration review feedback
